+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Date format: (YYYY-MM-DD)
   * Removed the importing of pre MC 1.14 cat typed ocelot shopkeepers.
   * Removed the importing of pre MC 1.14 villager profession data.
   * Removed the 'unknown' owner name validation (no longer used since late 1.14.4). This conflicted with any player actually named 'unknown'.
+* Fixed: Striders no longer randomly spawn with a saddle.
 
 **Debugging changes:**  
 * We log now when we unexpectedly encounter invalid currency items in the editors of the selling, buying, or book shopkeepers.